Various error handling improvments
